Rename sub-module with HCTL tree struct.
