+++ src/preprocessing/mod.rs
@@ -1,7 +1,7 @@
 //! Components regarding the preprocessing of HCTL formulae, e.g., tokenizing, parsing, validating,
 //! or renaming of variables.
 
-pub mod node;
+pub mod hctl_tree;
 pub mod operator_enums;
 pub mod parser;
 pub mod read_inputs;
